+++ source/client/benchmark_client_impl.h
@@ -48,7 +48,15 @@ struct BenchmarkClientStats {
 class Http1PoolImpl : public Envoy::Http::Http1::ProdConnPoolImpl {
 public:
   using Envoy::Http::Http1::ProdConnPoolImpl::ProdConnPoolImpl;
-  void createConnections(const uint32_t connection_limit);
+  Envoy::Http::ConnectionPool::Cancellable*
+  newStream(Envoy::Http::StreamDecoder& response_decoder,
+            Envoy::Http::ConnectionPool::Callbacks& callbacks) override;
+  void setPrefetchConnections(const bool prefetch_connections) {
+    prefetch_connections_ = prefetch_connections;
+  }
+
+private:
+  bool prefetch_connections_{};
 };
 
 // Vanilla Envoy's HTTP/2 pool is single connection only (or actually sometimes dual in connection
@@ -161,8 +169,6 @@ public:
     return cluster_manager_->httpConnPoolForCluster(
         cluster_name_, Envoy::Upstream::ResourcePriority::Default, proto, nullptr);
   }
-
-  void prefetchPoolConnections() override;
 
 private:
   Envoy::Api::Api& api_;